backend/src/requirements.txt
django-cors-headers
djangorestframework
djangorestframework-simplejwt
Pillow
requests
scipy
matplotlib
gunicorn